txsearch_tests.cpp initialized
